+++ cluster/cluster.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 
 	"github.com/chrislonng/starx/log"
-	"github.com/chrislonng/starx/network/rpc"
+	"github.com/chrislonng/starx/cluster/rpc"
 	"github.com/chrislonng/starx/session"
 )
