+++ network/handler.go
@@ -8,11 +8,11 @@ import (
 	"runtime"
 
 	"github.com/chrislonng/starx/log"
-	"github.com/chrislonng/starx/message"
 	"github.com/chrislonng/starx/network/route"
-	"github.com/chrislonng/starx/network/rpc"
-	"github.com/chrislonng/starx/packet"
+	"github.com/chrislonng/starx/cluster/rpc"
 	"github.com/chrislonng/starx/session"
+	"github.com/chrislonng/starx/network/message"
+	"github.com/chrislonng/starx/network/packet"
 )
 
 // Unhandled message buffer size
