+++ network/message/message.go
